Read multiple measurements per well

single_time_multiple_reads() gets a Tecan plate reader measurement from an Excel file when multiple readings of each well were taken

Usage

single_time_multiple_reads(dat_raw)

Arguments

dat_raw

(tibble) a excel sheet

Value

A tibble::tibble() containing tidy data

Examples

traw$single_time_multiple_reads |> single_time_multiple_reads()
#> # A tibble: 96 × 7
#>    Well    Mean   StDev  `1;2`  `0;1`  `1;0`  `2;1`
#>    <chr>  <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1 A1    0.104  0.0243  0.0960 0.0896 0.0891 0.140 
#>  2 A2    0.0951 0.00730 0.0913 0.0913 0.0918 0.106 
#>  3 A3    0.105  0.0279  0.0944 0.0885 0.0893 0.146 
#>  4 A4    0.102  0.0215  0.0933 0.0899 0.0907 0.134 
#>  5 A5    0.0950 0.0122  0.0887 0.0887 0.0891 0.113 
#>  6 A6    0.0933 0.00280 0.0944 0.0928 0.0897 0.0964
#>  7 A7    0.0968 0.00940 0.0932 0.0932 0.0900 0.111 
#>  8 A8    0.0941 0.00560 0.0914 0.0901 0.0925 0.102 
#>  9 A9    0.0945 0.00910 0.0898 0.0897 0.0903 0.108 
#> 10 A10   0.0935 0.00680 0.0901 0.0905 0.0897 0.104 
#> # ℹ 86 more rows