Get time series of temperature values

tread_temperature() was deprecated because it only supported one type of data (time series with multiple readings per well) and the same result can be achieved with tparse()

Usage

tread_temperature(xlsx_file, xlsx_sheet = 1)

Arguments

xlsx_file

(character) path to Excel file

xlsx_sheet

(numeric) index of Excel sheet to read (default: 1)

Value

A tibble::tibble() containing tidy data

Examples

tread_temperature(
  system.file(
    "extdata",
    "time_series_multiple_reads.xlsx",
    package = "tread"
  )
)
#> Warning: `tread_temperature()` was deprecated in tread 0.4.0.
#> ℹ This function is a special case of tparse(); use it instead
#> # A tibble: 25 × 2
#>     time temperature
#>    <int>       <dbl>
#>  1     0        30.6
#>  2   600        30.3
#>  3  1200        30.6
#>  4  1800        30.7
#>  5  2400        30.5
#>  6  3000        30.5
#>  7  3600        30.3
#>  8  4200        30.4
#>  9  4800        30.5
#> 10  5400        30.6
#> # ℹ 15 more rows
# ->
tparse(
  system.file(
    "extdata",
    "time_series_multiple_reads.xlsx",
    package = "tread"
  )
) |>
dplyr::select(time, temp) |>
dplyr::distinct()
#> ℹ Multiple reads per well detected
#> ℹ Time series detected
#> # A tibble: 20 × 2
#>      time  temp
#>     <dbl> <dbl>
#>  1     0   30.6
#>  2   600   30.3
#>  3  1200   30.6
#>  4  1800   30.7
#>  5  2400   30.5
#>  6  3000   30.5
#>  7  3600   30.3
#>  8  4200   30.4
#>  9  4800   30.5
#> 10  5400   30.6
#> 11  6000.  30.8
#> 12  6600.  30.8
#> 13  7200.  31  
#> 14  7800.  30.9
#> 15  8400.  30.9
#> 16  9000.  31.2
#> 17  9600.  31.2
#> 18 10200.  31.3
#> 19 10800.  31  
#> 20 11400.  31.2